requirements.txt
streamlit
firebase-admin
requests
python-dotenv
openpyxl
groq
pandas